+++ packages/playground/resources/js/app.js
@@ -2,7 +2,7 @@ import { createApp, h } from 'vue';
 import { createInertiaApp } from '@inertiajs/inertia-vue3';
 import { InertiaProgress } from '@inertiajs/progress';
 import { plugin as formkitPlugin, defaultConfig } from '@formkit/vue';
-import inertiaFormkitPlugin from 'inertiajs-formkit-plugin';
+import formkitInertify from 'formkit-inertify';
 
 import "../css/app.css";
 
@@ -12,7 +12,7 @@ createInertiaApp({
     createApp({ render: () => h(app, props) })
       .use(plugin)
       .use(formkitPlugin, defaultConfig({
-        plugins: [inertiaFormkitPlugin]
+        plugins: [formkitInertify]
       }))
       .mount(el);
   }