+++ packages/lib/CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 
-### 1.1.0
+### 2.0.0-alpha.1
+
+#### BREAKING CHANGE
+
+* Changed the packages name from `inertiajs-formkit-plugin` to `formkit-inertify`.
+* Removed `inertiaOptions` as third parameters to FormKit `@submit`.
+* Added to FormKit's `context` an `inertia` property to be used for visits.
+
+### 1.1.0 [DEPRECATED](https://www.npmjs.com/package/inertiajs-formkit-plugin)
 
 #### Features
 
 * Added a third parameter `inertiaOptions` to FormKit form `@submit`.
 
-### 1.0.0
+### 1.0.0 [DEPRECATED](https://www.npmjs.com/package/inertiajs-formkit-plugin)
 
 #### Features